requirements.txt
openai
langchain
faiss-cpu
pydantic
python-dotenv
numpy
langchain-community
rank-bm25